исправила 3 файл

+++ lesson_02/lesson_2_task_3.py
@@ -1,5 +1,5 @@
 import math
-num_items = int(input("Введите сторону квадрата: "))
+num_items = float(input("Введите сторону квадрата: "))
 
 def square(side):
     area = side * side
